Create README.md
**Hi there! **

I'm Ayush Sarathe, an 3rd year Engineering student with a strong passion for Computer Science and the ever-evolving world of the Internet of Things (IoT).

In this corner of the web, you'll find:

Projects: Explore my endeavors in computer science, with a focus on building and programming solutions for the IoT landscape.
Exploration: Witness my journey of learning and experimentation as I delve deeper into fascinating areas of CS and IoT.
Collaboration: I'm always eager to connect and collaborate on projects that push the boundaries of technology, especially within the realm of IoT.


Stay Connected:

https://www.linkedin.com/in/ayush-sarathe-20576523a
Feel free to reach out for any discussions or collaborations!
This is a work in progress!

I'm constantly learning and growing in my CS and IoT journey.  Be sure to check back for updates and exciting new projects!